fix: comprehensive code quality improvements across all 15 services
Phase 1: SubscriptionClient fail-open pattern + logging (8 clients)
Phase 2: Float → Numeric(12,2) for 15 monetary columns + 5 migrations
Phase 3: Error handling fixes in account/debt services
Phase 4: N+1 query fixes, ThreadPoolExecutor, async→sync, httpx reuse
Phase 5: CORS wildcard → configurable, bare except → Exception, traceback imports
Phase 6: pool_pre_ping/pool_recycle for 7 DBs, lifespan migration for 8 services
Phase 7: 204 deletes, version standardization, DEFAULT_CURRENCY, verbose logging

+++ account_service/app/models/account.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Float, DateTime, Boolean, Enum, Index
+from sqlalchemy import Column, Integer, String, Numeric, DateTime, Boolean, Enum, Index
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.sql import func
 from app.database import Base
@@ -20,7 +20,7 @@ class Account(Base):
     name = Column(String(100), nullable=False)
     type = Column(Enum(AccountType), nullable=False)
     currency = Column(String(3), nullable=False, default="USD")
-    balance = Column(Float, nullable=False, default=0.0)
+    balance = Column(Numeric(12, 2), nullable=False, default=0.0)
     description = Column(String(500), nullable=True)
     is_active = Column(Boolean, default=True, nullable=False)
     is_archived = Column(Boolean, default=False, nullable=False)
